downloadBinsWhenPossible: Get macos-13 bins, not macos-12 bins
Core Pulsar repo is building on macOS 13 CI runners now.

Ensure we get all the bins. Without this change, we skip downloading
the macOS bins, as the "macos-12 Binaries" string doesn't match
"macos-13 Binaries".

Might be more graceful solutions to this, but we can hotfix it this way
for now.